chore: remove unnecessary interface
