+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/export/trackedentity/aggregates/EnrollmentAggregate.java
@@ -28,6 +28,8 @@
 package org.hisp.dhis.tracker.export.trackedentity.aggregates;
 
 import static java.util.concurrent.CompletableFuture.allOf;
+import static org.hisp.dhis.tracker.export.trackedentity.aggregates.AsyncUtils.asyncFetch;
+import static org.hisp.dhis.tracker.export.trackedentity.aggregates.AsyncUtils.conditionalAsyncFetch;
 import static org.hisp.dhis.tracker.export.trackedentity.aggregates.ThreadPoolManager.getPool;
 
 import com.google.common.collect.Multimap;
@@ -51,7 +53,7 @@ import org.springframework.stereotype.Component;
  */
 @Component("org.hisp.dhis.tracker.trackedentity.aggregates.EnrollmentAggregate")
 @RequiredArgsConstructor
-public class EnrollmentAggregate implements Aggregate {
+class EnrollmentAggregate {
   @Qualifier("org.hisp.dhis.tracker.trackedentity.aggregates.EnrollmentStore")
   @Nonnull
   private final EnrollmentStore enrollmentStore;
