+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/export/trackedentity/aggregates/TrackedEntityStore.java
@@ -27,58 +27,170 @@
  */
 package org.hisp.dhis.tracker.export.trackedentity.aggregates;
 
+import com.google.common.collect.ArrayListMultimap;
+import com.google.common.collect.Lists;
 import com.google.common.collect.Multimap;
+import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import org.hisp.dhis.relationship.RelationshipItem;
+import org.apache.commons.lang3.StringUtils;
 import org.hisp.dhis.trackedentity.TrackedEntity;
 import org.hisp.dhis.trackedentity.TrackedEntityProgramOwner;
 import org.hisp.dhis.trackedentityattributevalue.TrackedEntityAttributeValue;
+import org.hisp.dhis.tracker.export.trackedentity.aggregates.mapper.OwnedTeMapper;
+import org.hisp.dhis.tracker.export.trackedentity.aggregates.mapper.ProgramOwnerRowCallbackHandler;
+import org.hisp.dhis.tracker.export.trackedentity.aggregates.mapper.TrackedEntityAttributeRowCallbackHandler;
+import org.hisp.dhis.tracker.export.trackedentity.aggregates.mapper.TrackedEntityRowCallbackHandler;
+import org.hisp.dhis.tracker.export.trackedentity.aggregates.query.TeAttributeQuery;
+import org.hisp.dhis.tracker.export.trackedentity.aggregates.query.TrackedEntityQuery;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.stereotype.Repository;
 
 /**
  * @author Luciano Fiandesio
+ * @author Ameen Mohamed
  */
-public interface TrackedEntityStore {
-  /**
-   * Get a Map of {@see TrackedEntity} by Primary Keys
-   *
-   * @param ids a list of Tracked Entity Primary Keys
-   * @return a Map where key is a {@see TrackedEntity} uid and the key is the corresponding {@see
-   *     TrackedEntity}
-   */
-  Map<String, TrackedEntity> getTrackedEntities(List<Long> ids, Context ctx);
+@Repository
+class TrackedEntityStore extends AbstractStore {
+  private static final String GET_TE_SQL = TrackedEntityQuery.getQuery();
 
-  /**
-   * Fetches all the relationships having the TE id specified in the arg as "left" or "right"
-   * relationship
-   *
-   * @param ids a list of Tracked Entity Primary Keys
-   * @return a MultiMap where key is a {@see TrackedEntity} uid and the key a List of {@see
-   *     Relationship} objects
-   */
-  Multimap<String, RelationshipItem> getRelationships(List<Long> ids, Context ctx);
+  private static final String GET_TE_ATTRIBUTES = TeAttributeQuery.getQuery();
 
-  /**
-   * @param ids @param ids a list of Tracked Entity Primary Keys
-   * @return a MultiMap where key is a {@see TrackedEntity} uid and the key a List of {@see
-   *     Attribute} objects
-   */
-  Multimap<String, TrackedEntityAttributeValue> getAttributes(List<Long> ids);
+  private static final String GET_PROGRAM_OWNERS =
+      "select te.uid as key, p.uid as prguid, o.uid as ouuid "
+          + "from trackedentityprogramowner teop "
+          + "join program p on teop.programid = p.programid "
+          + "join organisationunit o on teop.organisationunitid = o.organisationunitid "
+          + "join trackedentity te on teop.trackedentityid = te.trackedentityid "
+          + "where teop.trackedentityid in (:ids)";
 
-  /**
-   * @param ids a list of Tracked Entity Primary Keys
-   * @return a MultiMap where key is a {@see TrackedEntity} uid and the * key a List of {@see
-   *     ProgramOwner} objects
-   */
-  Multimap<String, TrackedEntityProgramOwner> getProgramOwners(List<Long> ids);
+  private static final String FILTER_OUT_DELETED_TE = "te.deleted=false";
 
-  /**
-   * For each te, get the list of programs for which the user has ownership.
-   *
-   * @param ids a list of Tracked Entity primary keys
-   * @param ctx aggregate context
-   * @return Tei uids mapped to a list of program uids to which user has ownership
-   */
-  Multimap<String, String> getOwnedTrackedEntities(
-      List<Long> ids, Context ctx, boolean skipUserScopeValidation);
+  private String getTrackedEntitiesOwnershipSqlForAllPrograms(boolean skipUserScopeValidation) {
+    String sql =
+        "SELECT te.uid as te_uid,tpo.trackedentityid, tpo.programid, tpo.organisationunitid, p.accesslevel,p.uid as pgm_uid "
+            + "FROM trackedentityprogramowner TPO "
+            + "LEFT JOIN program P on P.programid = TPO.programid "
+            + "LEFT JOIN organisationunit OU on OU.organisationunitid = TPO.organisationunitid "
+            + "LEFT JOIN trackedentity TE on TE.trackedentityid = tpo.trackedentityid "
+            + "WHERE TPO.trackedentityid in (:ids) "
+            + "AND p.programid in (SELECT programid FROM program) ";
+
+    if (!skipUserScopeValidation)
+      sql +=
+          "GROUP BY te.uid,tpo.trackedentityid, tpo.programid, tpo.organisationunitid, ou.path, p.accesslevel,p.uid "
+              + "HAVING (P.accesslevel in ('OPEN', 'AUDITED') AND (EXISTS(SELECT SS.organisationunitid FROM userteisearchorgunits SS LEFT JOIN organisationunit OU2 ON OU2.organisationunitid = SS.organisationunitid WHERE userinfoid = :userInfoId AND OU.path LIKE CONCAT(OU2.path, '%')) OR EXISTS(SELECT CS.organisationunitid FROM usermembership CS LEFT JOIN organisationunit OU2 ON OU2.organisationunitid = CS.organisationunitid WHERE userinfoid = :userInfoId AND OU.path LIKE CONCAT(OU2.path, '%')))) "
+              + "OR (P.accesslevel in ('CLOSED', 'PROTECTED') AND EXISTS(SELECT CS.organisationunitid FROM usermembership CS LEFT JOIN organisationunit OU2 ON OU2.organisationunitid = CS.organisationunitid WHERE userinfoid = :userInfoId AND OU.path LIKE CONCAT(OU2.path, '%')));";
+
+    return sql;
+  }
+
+  private String getTrackedEntitiesOwnershipSqlForSpecificProgram(boolean skipUserScopeValidation) {
+    String sql =
+        "SELECT te.uid as te_uid,tpo.trackedentityid, tpo.programid, tpo.organisationunitid, p.accesslevel,p.uid as pgm_uid "
+            + "FROM trackedentityprogramowner TPO "
+            + "LEFT JOIN program P on P.programid = TPO.programid "
+            + "LEFT JOIN organisationunit OU on OU.organisationunitid = TPO.organisationunitid "
+            + "LEFT JOIN trackedentity TE on TE.trackedentityid = tpo.trackedentityid "
+            + "WHERE TPO.trackedentityid in (:ids) "
+            + "AND p.uid = :programUid ";
+
+    if (!skipUserScopeValidation) {
+      sql +=
+          "GROUP BY te.uid,tpo.trackedentityid, tpo.programid, tpo.organisationunitid, ou.path, p.accesslevel,p.uid "
+              + "HAVING (P.accesslevel in ('OPEN', 'AUDITED') AND (EXISTS(SELECT SS.organisationunitid FROM userteisearchorgunits SS LEFT JOIN organisationunit OU2 ON OU2.organisationunitid = SS.organisationunitid WHERE userinfoid = :userInfoId AND OU.path LIKE CONCAT(OU2.path, '%')) OR EXISTS(SELECT CS.organisationunitid FROM usermembership CS LEFT JOIN organisationunit OU2 ON OU2.organisationunitid = CS.organisationunitid WHERE userinfoid = :userInfoId AND OU.path LIKE CONCAT(OU2.path, '%')))) "
+              + "OR (P.accesslevel in ('CLOSED', 'PROTECTED') AND EXISTS(SELECT CS.organisationunitid FROM usermembership CS LEFT JOIN organisationunit OU2 ON OU2.organisationunitid = CS.organisationunitid WHERE userinfoid = :userInfoId AND OU.path LIKE CONCAT(OU2.path, '%')));";
+    }
+
+    return sql;
+  }
+
+  public TrackedEntityStore(@Qualifier("readOnlyJdbcTemplate") JdbcTemplate jdbcTemplate) {
+    super(jdbcTemplate);
+  }
+
+  @Override
+  String getRelationshipEntityColumn() {
+    return "trackedentityid";
+  }
+
+  public Map<String, TrackedEntity> getTrackedEntities(List<Long> ids, Context ctx) {
+    List<List<Long>> idPartitions = Lists.partition(ids, PARITITION_SIZE);
+
+    Map<String, TrackedEntity> trackedEntityMap = new LinkedHashMap<>();
+
+    idPartitions.forEach(
+        partition -> trackedEntityMap.putAll(getTrackedEntitiesPartitioned(partition, ctx)));
+    return trackedEntityMap;
+  }
+
+  private Map<String, TrackedEntity> getTrackedEntitiesPartitioned(List<Long> ids, Context ctx) {
+    TrackedEntityRowCallbackHandler handler = new TrackedEntityRowCallbackHandler();
+
+    if (!ctx.isSuperUser() && ctx.getTrackedEntityTypes().isEmpty()) {
+      // If not super user and no tets are accessible. then simply return
+      // empty list.
+      return new HashMap<>();
+    }
+
+    String sql =
+        getQuery(GET_TE_SQL, ctx, "te.trackedentitytypeid in (:teTypeIds)", FILTER_OUT_DELETED_TE);
+    jdbcTemplate.query(
+        applySortOrder(sql, StringUtils.join(ids, ",")),
+        createIdsParam(ids).addValue("teTypeIds", ctx.getTrackedEntityTypes()),
+        handler);
+
+    return handler.getItems();
+  }
+
+  public Multimap<String, TrackedEntityAttributeValue> getAttributes(List<Long> ids) {
+    return fetch(GET_TE_ATTRIBUTES, new TrackedEntityAttributeRowCallbackHandler(), ids);
+  }
+
+  public Multimap<String, TrackedEntityProgramOwner> getProgramOwners(List<Long> ids) {
+    return fetch(GET_PROGRAM_OWNERS, new ProgramOwnerRowCallbackHandler(), ids);
+  }
+
+  public Multimap<String, String> getOwnedTrackedEntities(
+      List<Long> ids, Context ctx, boolean skipUserScopeValidation) {
+    List<List<Long>> teds = Lists.partition(ids, PARITITION_SIZE);
+
+    Multimap<String, String> ownedTeisMultiMap = ArrayListMultimap.create();
+
+    teds.forEach(
+        partition ->
+            ownedTeisMultiMap.putAll(
+                getOwnedTeisPartitioned(partition, ctx, skipUserScopeValidation)));
+
+    return ownedTeisMultiMap;
+  }
+
+  private Multimap<String, String> getOwnedTeisPartitioned(
+      List<Long> ids, Context ctx, boolean skipUserScopeValidation) {
+    OwnedTeMapper handler = new OwnedTeMapper();
+
+    MapSqlParameterSource paramSource = createIdsParam(ids).addValue("userInfoId", ctx.getUserId());
+
+    boolean checkForOwnership =
+        ctx.getParams().isIncludeEnrollments()
+            || ctx.getParams().getTeEnrollmentParams().isIncludeEvents();
+
+    String sql;
+
+    if (ctx.getQueryParams().hasProgram()) {
+      sql = getTrackedEntitiesOwnershipSqlForSpecificProgram(skipUserScopeValidation);
+      paramSource.addValue("programUid", ctx.getQueryParams().getProgram().getUid());
+    } else if (checkForOwnership) {
+      sql = getTrackedEntitiesOwnershipSqlForAllPrograms(skipUserScopeValidation);
+    } else {
+      return ArrayListMultimap.create();
+    }
+
+    jdbcTemplate.query(sql, paramSource, handler);
+
+    return handler.getItems();
+  }
 }
